feat(cargo): implement cargo feature with fake data
This commit introduces the cargo feature, including:
- `CargoRepository` interface and its implementation `CargoRepositoryImpl` using a `FakeDataSource`.
- `GetCargoUseCase` for retrieving cargo data.
- `CargoItemMapper` for converting between DTO and domain models.
- `HomeViewModel` to manage UI state for the cargo list.
- `HomeScreen` composable to display the list of cargo items.
- Updated `CargoItemDto` to use `Long` for `priceToman`.
- Removed `.gitkeep` files from now populated directories.

+++ app/src/main/java/com/example/barflowapp/data/model/CargoItemDto.kt
@@ -16,7 +16,7 @@ data class CargoItemDto(
     @SerializedName("weight_ton")
     val weightTon: Double,
     @SerializedName("price_toman")
-    val priceToman: Int,
+    val priceToman: Long,
     @SerializedName("cargo_type")
     val cargoType: String,
     @SerializedName("packaging")
